Fix template code for web app toolbox
When Jekyll renders the inline code for the template it tries to
interpret the {{ name }} section, leaving an empty string on the
processed toolbox web page. Additionally, the if/else/endif
statements were missing the {% %} syntax.
I fixed the if statements and wrapped the entire code section with
{% raw %} {% endraw %} in order to display correctly when processed
by Jekyll (note that they will show up in GitHub's markdown
preview).